Implements base of SystemPickerFragment

+++ app/src/main/java/com/samsung/hyunjaee/scopedstorage/domain/entity/Media.kt
@@ -2,9 +2,8 @@ package com.samsung.hyunjaee.scopedstorage.domain.entity
 
 import android.net.Uri
 
-
 data class Media(
     val uri: Uri,
     val name: String,
-    val size: Int
+    val size: Long
 )
